refactor: add split search controller and view
